Addresses PR's comments: Move Batch_v1 to model.v1.  Remove unused code in Value
